% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stereo3D.R
\name{Stereo3D}
\alias{Stereo3D}
\title{Create Stereo3D image of given data}
\usage{
Stereo3D(data_file, stereo_angle, distance, connection_file )
}
\arguments{
\item{data_file}{A tab seperated file with \emph{.tsv} extension and having five columns \emph{(index, X, Y, Z and Color)} of the data. Where, \emph{X, Y and Z} represent cordinates of a datapoint, \emph{Color} is the label of the given data point and \emph{index} clolumn have the index information of the datapoints.}

\item{stereo_angle}{angle by which 3D data to be rotated along Y-axis. \emph{Default:} 5 degree.}

\item{distance}{Gap between stereo image and original image.}

\item{connection_file}{A tab seperated file (optional) . Where, \emph{first} and \emph{second} column has indices of start and end point of a connection respectively.}
}
\value{
Two Stereo3D images
\enumerate{
 \item  Create Stereoscopic 3D plot with input data filename prefix and \strong{_Stereo.pdf} extention.
 \item  Interactive 3D plot for the above image pops up, which can be zoomed and rotated by draging the mouse.
 }
}
\description{
Stereo3D: Using stereo images to enrich 3D visualization
}
\examples{
connection_fileName=system.file("extdata", "connection_file.tsv", package = "Stereo3D", mustWork = TRUE)
sample_data_file=system.file("extdata", "sample_3D_data.tsv", package = "Stereo3D", mustWork = TRUE)
Stereo3D(data_file=sample_data_file, stereo_angle=5, distance=0, connection_file=connection_fileName)
}
\keyword{3D}
\keyword{Stereoscopic}
\keyword{plot}
\keyword{plot,}
\keyword{stereo}
